moved some data structures to separate files
